Reconsider coinbase witness commitment
Rather than having the coinbase witness commitment be the hash of the
concatenation of the transaction witness merkle root and a single 32-byte
nonce from the coinbase witness, treat it instead as the merkle root
of a tree with the leaves being the transaction witness merkle root,
and each of the items from the coinbase stack.

For the case where the coinbase witness stack contains a single, 32-byte
item this change has no effect, however it should generalise well to allow
the coinbase witness stack to contain an arbitrary number of elements.

(If an element on the coinbase stack is not already exactly 32 bytes,
it is first hashed before being added as a leaf to the merkle tree)