net: disentangle GetMessageToSend label from message type
This is in preparation for making Transports message type-agnostic. After
this change, Transport users can set the label for a message to send to be
any string, which will be echoed back in GetMessageToSend. The code in net,
and all current tests, keep setting the label equal to the message type, but
this is not mandatory from the perspective of Transport.