txgraph: stop using PostLinearize to guarantee connected chunks
Drop the expectation that txgraph chunks will always be connected. This used
to be guaranteed by running PostLinearize on the linearization output in case
it didn't report an optimal result, which was justified by the fact that
PostLinearize had, in the worst case, a much higher improvement-per-time rate
than the old search-based linearization algorithm.

With SFL, it seems better to instead use the allotted CPU time to perform more
SFL optimization steps (which might help finding the optimal) rather than
using it to just guarantee connected chunks.