clusterlin: do not attempt merge down sequence upon self-merge
After splitting a chunk, and performing a merge sequence on the top chunk to
make it topological again, it is possible that it merges back with the
bottom. When such a self-merge happens, it is pointless to attempt a merge
sequence on that same bottom again.

This is also a preparation for a later commit that keeps track of these
self-merges.